Rendering of product groups
We can now render these Product Groups but it is very obvious that we
have some data loading issues.  It seems Ember Data doesn't fetch (or
update) the relationships on first load of the content (or some other
mishap).  Anyhow, the products aren't visible when viewing the data from
Ember Data directly.