Abstraction for creating links
Creating links is done by manually supplying the target route and the
desired label.  This is error-prone and not extremely clear.  For
consistency, it would be nicer if we could just point to a resource and
say "link there, bro" and have the linking be done with.

We have done so in this refactoring.  Some further cleanup to the API is
possible.