Allow property editing based on input type
The input type of a property can provide us with a default entry on how
to edit the property.  This could make it easier to construct the
displays for manipulating various properties.

We've implemented a basic system here that could be extended.  Ideally,
this system is extended further so the manipulators can live in their
own files and can so they can also be tailored for specific
relationships.