A more-or-less generic page header
This page header will generate links to other relevant pages based on
the information available in the model.

Some extended work could happen on this component as we may be
interested in rendering links only if the supplied data is
available (eg: some pages may not have edit links or may not have show
links).