Allow adding business entities and setting them for locations
A location can now have a businessEntity connected to it.  This is tackled
through the supplier relation.

It is possible to set the businessEntity for the location now such that we can
later learn to filter products on them.

Although a BusinessEntity could already be manipulated through the suppliers,
there was no interface to create new ones yet.  This has been added also.